fix: prevent users from changing another user's data

+++ backend/src/users/users.controller.ts
@@ -7,6 +7,8 @@ import {
   Body,
   UseGuards,
   Param,
+  Req,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { AuthGuard } from '../auth/guards/auth.guard';
@@ -18,28 +20,37 @@ export class UsersController {
 
   @HttpCode(HttpStatus.OK)
   @Put(':id/name')
-  async updateName(@Param('id') id: string, @Body() input: { name: string }) {
+  async updateName(@Param('id') id: string, @Body() input: { name: string }, @Req() req: any) {
+    if (req.user.userId !== Number(id)) {
+      throw new ForbiddenException('You are not allowed to update another user\'s data.');
+    }
     return this.usersService.updateUserName(Number(id), input.name);
   }
 
   @HttpCode(HttpStatus.OK)
   @Put(':id/email')
-  async updateEmail(@Param('id') id: string, @Body() input: { email: string }) {
+  async updateEmail(@Param('id') id: string, @Body() input: { email: string }, @Req() req: any) {
+    if (req.user.userId !== Number(id)) {
+      throw new ForbiddenException('You are not allowed to update another user\'s data.');
+    }
     return this.usersService.updateUserEmail(Number(id), input.email);
   }
 
   @HttpCode(HttpStatus.OK)
   @Put(':id/password')
-  async updatePassword(
-    @Param('id') id: string,
-    @Body() input: { password: string },
-  ) {
+  async updatePassword(@Param('id') id: string, @Body() input: { password: string }, @Req() req: any) {
+    if (req.user.userId !== Number(id)) {
+      throw new ForbiddenException('You are not allowed to update another user\'s data.');
+    }
     return this.usersService.updateUserPassword(Number(id), input.password);
   }
 
   @HttpCode(HttpStatus.NO_CONTENT)
   @Delete(':id')
-  async deleteUser(@Param('id') id: string) {
+  async deleteUser(@Param('id') id: string, @Req() req: any) {
+    if (req.user.userId !== Number(id)) {
+      throw new ForbiddenException('You are not allowed to delete another user\'s data.');
+    }
     return this.usersService.deleteUser(Number(id));
   }
 }
